+++ Docker_R/build.sh
@@ -2,7 +2,7 @@
 #docker-compose -f docker-compose.yml build seurat
 #docker-compose -f docker-compose.yml build r
 
-tag=3.0.0
+tag=3.1.0
 for name in shortcake_seurat shortcake_r
 do
   docker push rnakato/$name:$tag